jestSample/functions.ts
import { NameApiService } from "./nameApiService";
import { DatabaseMock } from "./util";

export const sumOfArray = (numbers: number[]): number => {
  if (numbers.length) {
    return numbers.reduce((a: number, b: number): number => a + b);
  } else {
    return 0;
  }
};

export const asyncSumOfArray = (numbers: number[]): Promise<number> => {
  return new Promise((resolve): void => {
    resolve(sumOfArray(numbers));
  });
};

export const asyncSumOfArraySometimesZero = (
  numbers: number[],
  DatabaseMock: DatabaseMock
): Promise<number> => {
  return new Promise((resolve): void => {
    try {
      const database = DatabaseMock; // fixme: この関数をテストするには、DatabaseMockの使い方を変える必要がありそう！ヒント：依存性の注入
      database.save(numbers);
      resolve(sumOfArray(numbers));
    } catch (error) {
      resolve(0);
    }
  });
};

export const getFirstNameThrowIfLong = async (
  maxNameLength: number,
  nameApi: NameApiService,
  data: { [key: string]: string }
): Promise<string> => {
  const nameApiSerivce = nameApi; // fixme: この関数をテストするには、NameApiServiceの使い方を変える必要がありそう！ヒント：依存性の注入
  const firstName = await nameApiSerivce.getFirstName(data);

  if (firstName.length > maxNameLength) {
    throw new Error("first_name too long");
  }
  return firstName;
};
